+++ Runtime/TTType/PowerOfTwoAttribute.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace net.rs64.TexTransTool.Utils
+{
+    internal class PowerOfTwoAttribute : PropertyAttribute
+    {
+    }
+}